requirements.txt
Django==3.2.20
drf-yasg==1.21.5
djangorestframework==3.14.0
django-model-utils==4.3.1
psycopg2-binary==2.9.6
gunicorn>=19.9.0,<20
coverage[toml]==7.2.5
requests==2.29.0
requests-mock==1.9.3
django-filter==23.2
django-qr-code==3.1.0
drf-yasg==1.21.5
django-oidc-provider==0.7
PyJWT==2.4.0
cryptography==40.0.2
marshmallow==3.17.0
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
freezegun==1.2.2
django-sslserver==0.22
django-cors-headers==4.1.0
